show related images in the reuse tab of image-detail section
The user can also bookmark images from this image grid.
