+++ src/popup/popup.html
@@ -277,6 +277,13 @@
 
           <h5 class="b-header margin-top-bigger margin-bottom-normal">Tags</h5>
           <div class="image-tags"></div>
+
+          <h5 class="b-header margin-top-bigger margin-bottom-normal">Related Images</h5>
+          <div class="section-content__images-container">
+            <div class="grid grid-related-images">
+              <div class="gutter-sizer"></div>
+            </div>
+          </div>
         </div>
         <!-- Information Tab -->
         <div class="tabs-panel container information-panel" data-content-no="2">
